+++ src/Test.Automated/SharedCoreUnitCoverageSuite.cs
@@ -25,6 +25,7 @@ namespace Test.Automated
             await ExecuteTestsAsync(SharedMiddlewarePipelineTests.GetTests()).ConfigureAwait(false);
             await ExecuteTestsAsync(SharedWebSocketTests.GetTests()).ConfigureAwait(false);
             await ExecuteTestsAsync(SharedNetstandard21CompatTests.GetTests()).ConfigureAwait(false);
+            await ExecuteTestsAsync(SharedOpenApiCompositionTests.GetTests()).ConfigureAwait(false);
 
             return _Results.ToArray();
         }